Refactoring fake component and shipper codebases
