+++ pkg/component/fake/component/comp/dialer.go
@@ -5,7 +5,7 @@
 //go:build !windows
 // +build !windows
 
-package main
+package comp
 
 import (
 	"context"
