+++ pkg/component/fake/component/comp/dialer_windows.go
@@ -5,7 +5,7 @@
 //go:build windows
 // +build windows
 
-package main
+package comp
 
 import (
 	"context"
